requirements.txt
dynaconf>=3.2
fastapi>=0.112.1
Jinja2>=3.1.4
pillow>=10.4.0
pydantic>=2.8.2
uvicorn>=0.30.6
watchdog>=4.0.2
